requirements.txt
absl-py==1.0.0
albumentations==0.3.2
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dill @ file:///tmp/build/80754af9/dill_1623919422540/work
docker-pycreds==0.4.0
dominate==2.6.0
einops==0.5.0
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.29.1
fsspec==2022.3.0
future==0.18.2
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.16.1
imgaug==0.2.6
importlib-metadata==4.11.2
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mlconfig==0.1.4
monai==0.7.0
netron==5.7.4
networkx==2.6.3
nibabel==4.0.2
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.4
pytorch-lightning==1.1.0
pytz==2022.1
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.6
shortuuid==1.0.8
SimpleITK==2.2.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
thop==0.1.0.post2207010342
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.0+cu102
torchaudio==0.10.0+cu113
torchfile==0.1.0
torchstat==0.0.7
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.11.0+cu102
tornado==6.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
visdom==0.1.8.9
wandb==0.12.10
websocket-client==1.3.1
Werkzeug==2.1.1
wincertstore==0.2
yaspin==2.1.0
zipp==3.7.0
